+++ include/xb2at/readers/mesh_reader.h
@@ -1,8 +1,8 @@
 #pragma once
-#include <core.h>
+#include <xb2at/core.h>
 #include <modeco/Logger.h>
 
-#include <structs/mesh.h>
+#include <xb2at/structs/mesh.h>
 
 namespace xb2at {
 namespace core {
