+++ include/xb2at/readers/mibl_reader.h
@@ -1,9 +1,9 @@
 #pragma once
-#include <core.h>
+#include <xb2at/core.h>
 #include <modeco/Logger.h>
 
-#include <structs/mibl.h>
-#include <structs/xbc1.h>
+#include <xb2at/structs/mibl.h>
+#include <xb2at/structs/xbc1.h>
 
 namespace xb2at {
 namespace core {
