+++ include/xb2at/serializers/MIBLDeswizzler.h
@@ -1,7 +1,7 @@
 #pragma once
-#include <core.h>
+#include <xb2at/core.h>
 #include <modeco/Logger.h>
-#include <structs/mibl.h>
+#include <xb2at/structs/mibl.h>
 
 namespace xb2at {
 	namespace core {
