+++ include/xb2at/streamhelper.h
@@ -1,5 +1,5 @@
 #pragma once
-#include <core.h>
+#include <xb2at/core.h>
 #include <cstdint>
 #include <iostream>
 #include <vector>
