+++ include/xb2at/structs/mesh.h
@@ -3,7 +3,7 @@
  * Mesh structures.
  */
 #pragma once
-#include <core.h>
+#include <xb2at/core.h>
 
 namespace xb2at {
 	namespace core {
