+++ include/xb2at/structs/mibl.h
@@ -3,7 +3,7 @@
  * MIBL (Multi IBL Texture) structures
  */
 #pragma once
-#include <core.h>
+#include <xb2at/core.h>
 
 namespace xb2at {
 namespace core {
