+++ include/xb2at/structs/msmd.h
@@ -1,5 +1,5 @@
 #pragma once
-#include <core.h>
+#include <xb2at/core.h>
 
 namespace xb2at {
