+++ include/xb2at/structs/mxmd.h
@@ -3,7 +3,7 @@
  * MXMD structures.
  */
 #pragma once
-#include <core.h>
+#include <xb2at/core.h>
 
 namespace xb2at {
 	namespace core {
