+++ include/xb2at/structs/xbc1.h
@@ -3,7 +3,7 @@
  * XBC1 structures.
  */
 #pragma once
-#include <core.h>
+#include <xb2at/core.h>
 
 namespace xb2at {
 	namespace core {
