+++ src/core/serializers/model_serializer.cpp
@@ -1,4 +1,5 @@
-#include <serializers/model_serializer.h>
+#include <xb2at/serializers/model_serializer.h>
+
 #include <fx/gltf.h>
 #include <glm/mat4x4.hpp>
 #include <glm/common.hpp>
